third_party/WebKit/LayoutTests/html/dialog/dialog-close-event-expected.txt

Test that dialog receives a close event upon closing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dialog.open is false
PASS closedCount is 0
PASS selfDialog is dialog
PASS dialog.open is false
PASS event.cancelable is false
PASS closedCount is 1
PASS selfDialog is dialog
PASS dialog.open is false
PASS event.cancelable is false
PASS successfullyParsed is true

TEST COMPLETE
